requirements.txt
streamlit==1.24.0
numpy==1.24.3
pandas==2.0.3
matplotlib==3.7.1
plotly==5.18.0
pdfplumber==0.10.3
scikit-learn==1.3.0
openai==0.28.0
python-dotenv==1.0.0
pdfminer.six==20221105
jinja2==3.1.2
pdfkit==1.0.0
reportlab==4.0.7
PyPDF2==3.0.1
PyMuPDF==1.23.8
Pillow==10.1.0
pyperclip